add script argument parsing
